Improve request code
Use ClientHttpRequestInterceptor to add headers to the request which
removes the need for calling Jackson directly in order to use the
serialized body to sign the request.

Consolidate logic for creating the nonce and computing the
"Authorization" header in Credentials class.

Replace RestTemplate and Credentials beans with just PayeezyClient.

Rename PayeezyRequest -> PayeezyClient.

Externalize Payeezy transactions URL to applications.properties.